+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.13.0] - 2022-02-07
+
+### Added
+
+- Add page explaining Canada's optional holidays
+  - Add new page to sitemap
+- Added `Abbr` component with a nicer abbreviation experience
+- Added a little toast notification soliciting feedback
+
+### Updated
+
+- Add optional holidays to Alberta (for current year)
+
 ## [3.12.1] - 2022-02-06
 
 ### Updated